reordered commands, user output polish

+++ install.sh
@@ -1,21 +1,31 @@
 #!/usr/bin/env bash
-set -exo pipefail
-
-host="$1"
-IFS="" read -r -p "Enter license key: " -s key
-if [[ ! "$key" =~ ^[0-9a-zA-Z]{64}$ ]]; then
-    >&2 echo "Invalid license key."
-    exit 1
-fi
+set -eo pipefail
 
 dockerComposeFile="docker-compose.yml"
 dockerComposeFileCommit="3a4899b6f6d2cd487469f3c6b8dc10d84e664b22"
 dockerComposeFileSha256="175bcfd2bfc651563781451140946b81bbb5dc1f8f71ba992b1a259e72b4bebf"
 if [[ -f "$dockerComposeFile" ]]; then
     >&2 echo "$dockerComposeFile already exists."
-    exit 2
+    exit 1
+fi
+if [[ -f /etc/nginx/sites-enabled/everytrade ]]; then 
+    >&2 echo "/etc/nginx/sites-enabled/everytrade already exists."
+    exit 1
+fi
+if [[ -f /etc/nginx/sites-available/everytrade ]]; then 
+    >&2 echo "/etc/nginx/sites-available/everytrade already exists."
+    exit 1
+fi
+IFS="" read -r -p "Enter license key: " key
+if [[ ! "$key" =~ ^[0-9a-zA-Z]{64}$ ]]; then
+    >&2 echo "Invalid license key."
+    exit 1
 fi
 
+host="$1"
+if [[ -z "$host" ]]; then
+    host=$(ip route get 8.8.8.8 | head -1 | awk '{print $7}')
+fi
 username="user-${key:0:32}"
 password="${key:32:32}"
 
@@ -23,15 +33,10 @@ sudo apt-get update
 sudo apt-get -y install docker.io docker-compose nginx certbot python3-certbot-nginx
 #sudo usermod -a -G docker "$USER"
 
-echo "$password" | sudo docker login -u "$username" --password-stdin registry.everytrade.io
-
 curl "https://raw.githubusercontent.com/everytrade-io/everytrade-install/$dockerComposeFileCommit/docker-compose.yml" -o "$dockerComposeFile"
 sha256sum "$dockerComposeFile" | grep "$dockerComposeFileSha256"
+echo "$password" | sudo docker login -u "$username" --password-stdin registry.everytrade.io
 sudo docker-compose pull
-
-if [[ -z "$host" ]]; then
-    host=$(ip route get 8.8.8.8 | head -1 | awk '{print $7}')
-fi
 sudo EVERYTRADE_INSTALL_HOST="$host" docker-compose up -d
 
 sudo tee /etc/nginx/sites-available/everytrade > /dev/null <<EOF
@@ -61,6 +66,16 @@ server {
         }
 }
 EOF
-sudo rm /etc/nginx/sites-enabled/default
+if [[ -f /etc/nginx/sites-enabled/default ]]; then
+    sudo rm /etc/nginx/sites-enabled/default
+fi
 sudo ln -sf /etc/nginx/sites-available/everytrade /etc/nginx/sites-enabled/everytrade
 sudo systemctl reload nginx.service
+
+echo
+echo
+echo "Installation finished successfully."
+echo
+echo "Application is initializing exchange rate data. In a little while it'll be ready."
+echo -e "Open \033[0;33mhttp://$host/\033[0m in your browser to continue."
+echo
